clean base project using reset-project command
